+++ core/src/main/scala/ai/lum/odinson/utils/exceptions/OdinsonException.scala
@@ -11,7 +11,7 @@ class OdinsonException(message: String = null, cause: Throwable = null)
 object OdinsonException {
 
   val METADATA_STRING_FIELD_EXCEPTION: OdinsonException = new OdinsonException(
-    "StringField metadata is not supported.  Supported types: TokensField, NumberField, DateField"
+    "StringField metadata is not supported.  Supported types: TokensField, NumberField, DateField, NestedField"
   )
 
   def defaultMessage(message: String, cause: Throwable) =